Addressed comments and replaced sync interface with template interface
