+++ src/main/templates/io/lettuce/core/api/RedisBloomFilterCommands.java
@@ -1,0 +1,169 @@
+/*
+ * Copyright 2026-Present, Redis Ltd. and Contributors
+ * All rights reserved.
+ *
+ * Licensed under the MIT License.
+ *
+ * This file contains contributions from third-party contributors
+ * licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.lettuce.core.api;
+
+import java.util.List;
+
+import io.lettuce.core.bf.BfInfoValue;
+import io.lettuce.core.bf.arguments.BfInfoType;
+import io.lettuce.core.bf.arguments.BfInsertArgs;
+import io.lettuce.core.bf.arguments.BfReserveArgs;
+import io.lettuce.core.bf.BfScanDumpValue;
+
+/**
+ * ${intent} for Bloom Filter.
+ *
+ * @author Yordan Tsintsov
+ * @param <K> Key type.
+ * @param <V> Value type.
+ * @see <a href="https://redis.io/docs/latest/develop/data-types/probabilistic/bloom-filter/">Redis Bloom Filter</a>
+ * @since 7.6
+ */
+public interface RedisBloomFilterCommands<K, V> {
+
+    /**
+     * Add an item to the Bloom Filter.
+     *
+     * @param key the key.
+     * @param value the value.
+     * @return Boolean integer-reply {@code true} if the item was added, {@code false} if it was already in the filter.
+     */
+    Boolean bfAdd(K key, V value);
+
+    /**
+     * Get the number of items added and detected as unique in the Bloom Filter.
+     *
+     * @param key the key.
+     * @return Long integer-reply the number of items in the filter.
+     */
+    Long bfCard(K key);
+
+    /**
+     * Check if an item exists in the Bloom Filter.
+     *
+     * @param key the key.
+     * @param value the value.
+     * @return Boolean integer-reply {@code true} if the item may exist in the filter, {@code false} if it definitely does not.
+     */
+    Boolean bfExists(K key, V value);
+
+    /**
+     * Get information about the Bloom Filter.
+     *
+     * @param key the key.
+     * @return BfInfoValue the information about the filter.
+     */
+    BfInfoValue bfInfo(K key);
+
+    /**
+     * Get information about the Bloom Filter.
+     *
+     * @param key the key.
+     * @param type the type of information to get.
+     * @return BfInfoValue the information about the filter.
+     */
+    BfInfoValue bfInfo(K key, BfInfoType type);
+
+    /**
+     * Add one or more items to the Bloom Filter.
+     *
+     * @param key the key.
+     * @param values the values.
+     * @return a boolean list where {@code true} means that the item was added and {@code false} means that there's a
+     *      probability that the item was already added to the filter of the filter is full.
+     */
+    @SuppressWarnings("unchecked")
+    List<Boolean> bfInsert(K key, V... values);
+
+    /**
+     * Add one or more items to the Bloom Filter.
+     *
+     * @param key the key.
+     * @param insertArgs the insert arguments.
+     * @param values the values.
+     * @return a boolean list where {@code true} means that the item was added and {@code false} means that there's a
+     *      probability that the item was already added to the filter of the filter is full.
+     */
+    @SuppressWarnings("unchecked")
+    List<Boolean> bfInsert(K key, BfInsertArgs insertArgs, V... values);
+
+    /**
+     * Restores a Bloom filter previously saved using BF.SCANDUMP.
+     *
+     * @param key the key.
+     * @param iterator the iterator.
+     * @param data the data.
+     * @return String simple-string-reply {@code OK} if {@code BF.LOADCHUNK} was executed correctly.
+     */
+    String bfLoadChunk(K key, long iterator, byte[] data);
+
+    /**
+     * Add one or more items to the Bloom Filter.
+     *
+     * @param key the key.
+     * @param values the values.
+     * @return a boolean list of {@code true} or {@code false} where true means that the item was newly added and false means that
+     *      there's a probability that the item was already added to the filter.
+     */
+    @SuppressWarnings("unchecked")
+    List<Boolean> bfMAdd(K key, V... values);
+
+    /**
+     * Check if one or more items exist in the filter.
+     *
+     * @param key the key.
+     * @param values the values.
+     * @return a boolean list of {@code true} or {@code false} where true means that, with high probability, the item was
+     *      already added to the filter, and false means that key does not exist or that item was definitely not added to the filter.
+     */
+    @SuppressWarnings("unchecked")
+    List<Boolean> bfMExists(K key, V... values);
+
+    /**
+     * Creates an empty Bloom filter with a single sub-filter for the initial specified capacity and with an upper bound error rate.
+     *
+     * @param key the key.
+     * @param errorRate the error rate.
+     * @param capacity the capacity.
+     * @return String simple-string-reply {@code OK} if {@code BF.RESERVE} was executed correctly.
+     */
+    String bfReserve(K key, double errorRate, long capacity);
+
+    /**
+     * Creates an empty Bloom filter with a single sub-filter for the initial specified capacity and with an upper bound error rate.
+     *
+     * @param key the key.
+     * @param errorRate the error rate.
+     * @param capacity the capacity.
+     * @param reserveArgs the reserve arguments.
+     * @return String simple-string-reply {@code OK} if {@code BF.RESERVE} was executed correctly.
+     */
+    String bfReserve(K key, double errorRate, long capacity, BfReserveArgs reserveArgs);
+
+    /**
+     * Begins an incremental save of the Bloom filter.
+     *
+     * @param key the key.
+     * @param iterator the iterator.
+     * @return BfScanDumpValue the scan dump value.
+     */
+    BfScanDumpValue bfScanDump(K key, long iterator);
+
+}
