+++ src/test/java/io/lettuce/apigenerator/Constants.java
@@ -31,7 +31,7 @@ class Constants {
             "RedisGeoCommands", "RedisHashCommands", "RedisHLLCommands", "RedisKeyCommands", "RedisListCommands",
             "RedisScriptingCommands", "RedisSentinelCommands", "RedisServerCommands", "RedisSetCommands",
             "RedisSortedSetCommands", "RedisStreamCommands", "RedisStringCommands", "RedisTransactionalCommands",
-            "RedisJsonCommands", "RedisVectorSetCommands", "RediSearchCommands" };
+            "RedisJsonCommands", "RedisVectorSetCommands", "RediSearchCommands", "RedisBloomFilterCommands" };
 
     public static final File TEMPLATES = new File("src/main/templates");
